tests-env-ai-engine/requirements.txt
python-dotenv
langchain
huggingface_hub
langchain_community
transformers
torch
boto3
langchain_aws
strands-agents 
strands-agents-tools
strip-markdown
fastapi
uvicorn